Allow all inputs to be alwaysOn filters
Closes #1778
