+++ packages/ra-ui-materialui/src/list/FilterFormInput.js
@@ -1,0 +1,50 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Field } from 'redux-form';
+import IconButton from 'material-ui/IconButton';
+import ActionHide from '@material-ui/icons/HighlightOff';
+import classnames from 'classnames';
+import { translate } from 'ra-core';
+
+const emptyRecord = {};
+
+const sanitizeRestProps = ({ alwaysOn, ...props }) => props;
+
+const FilterFormInput = ({ filterElement, classes, resource, translate }) => (
+    <div
+        data-source={filterElement.props.source}
+        className={classnames('filter-field', classes.body)}
+    >
+        {filterElement.props.alwaysOn ? (
+            <div className={classes.spacer}>&nbsp;</div>
+        ) : (
+            <IconButton
+                className="hide-filter"
+                onClick={this.handleHide}
+                data-key={filterElement.props.source}
+                tooltip={translate('ra.action.remove_filter')}
+            >
+                <ActionHide />
+            </IconButton>
+        )}
+        <div>
+            <Field
+                allowEmpty
+                {...sanitizeRestProps(filterElement.props)}
+                name={filterElement.props.source}
+                component={filterElement.type}
+                resource={resource}
+                record={emptyRecord}
+            />
+        </div>
+    </div>
+);
+
+FilterFormInput.propTypes = {
+    filterElement: PropTypes.node,
+    classes: PropTypes.object,
+    resource: PropTypes.string,
+    translate: PropTypes.func,
+};
+
+export default translate(FilterFormInput);
